Refactor to reduce the complexity of the script
