+++ scripts/enrich_mobileapp_mappings/constants.py
@@ -1,0 +1,2 @@
+MAPPING_FIELDS = ("owasp_mastg", "owasp_mastg_know", "owasp_mastg_best", "owasp_maswe")
+MAPPING_FIELDS = ("owasp_mastg", "owasp_mastg_know", "owasp_mastg_best", "owasp_maswe")
